CC/5/auth-system/requirement.txt

Flask==2.0.1
firebase-admin==5.3.0
google-cloud-datastore==2.7.0
pyrebase4==4.5.0